\name{CarbonInmobilisationClass-class}
\Rdversion{1.1}
\docType{class}
\alias{CarbonInmobilisationClass-class}

\title{Class \code{"CarbonInmobilisationClass"}}
\description{
Stores data about carbon inmobilisation
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("CarbonInmobilisationClass", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{DINDEX}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{CIMMO1}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{CIMMO2}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{CIMMO3}:}{Object of class \code{"numeric"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "CarbonInmobilisationClass" in the signature.
}
\references{
%%  ~~put references to the literature/web site here~~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
}
\examples{
showClass("CarbonInmobilisationClass")
}
\keyword{classes}
