\name{CarbonMineralisationClass-class}
\Rdversion{1.1}
\docType{class}
\alias{CarbonMineralisationClass-class}

\title{Class \code{"CarbonMineralisationClass"}}
\description{
Stores data about carbon mineralisation
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("CarbonMineralisationClass", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{CA1DEC}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{CA2DEC}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{CA3DEC}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{CDOC1}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{CDOC2}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{CDOC3}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{CL1DEC}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{CL2DEC}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{CL3DEC}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{LI1DEC}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{LI2DEC}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{LI3DEC}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{OM1CAS}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{OM2CAS}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{OM3CAS}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{SOCNT1}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{SOCNT2}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{SOCNT3}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{MBSOL}:}{Object of class \code{"numeric"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "CarbonMineralisationClass" in the signature.
}
\references{
%%  ~~put references to the literature/web site here~~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
}
\examples{
showClass("CarbonMineralisationClass")
}
\keyword{classes}
