\name{ClimateClass-class}
\Rdversion{1.1}
\docType{class}
\alias{ClimateClass-class}

\title{Class \code{"ClimateClass"}}
\description{
Sets atmospheric CO2 level
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("ClimateClass", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{CO2}:}{Object of class \code{"numeric"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "ClimateClass" in the signature.
}
\references{
%%  ~~put references to the literature/web site here~~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
}
\examples{
showClass("ClimateClass")
}
\keyword{classes}
