\name{InfoCropR-package}
\alias{InfoCropR-package}
\alias{InfoCropR}
\docType{package}
\title{
InfoCropR, a modular crop model in R
}
\description{
InfoCropR is an R port of the InfoCrop model, originally written in FST, a FORTRAN based simulation language.
}
\details{
\tabular{ll}{
Package: \tab InfoCropR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-19\cr
License: \tab GPL (>=3)\cr
Depends: \tab methods\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Francisco Alvarez and Jacob van Etten

Maintainer: Jacob van Etten <jacobvanetten@yahoo.com>

}
\references{
D.W.G. van Kraalingen, C. Rappoldt & H.H. van Laar. 1994. "Appendix 5 The Fortran Simulation Translator (FST), a simulation language." In: J. Goudriaan & H.H. van Laar, 1994. Modelling potential crop growth processes. Kluwer Academic Publishers, Dordrecht. pp. 219-236.

D.W.G van Kraalingen, C Rappoldt, H.H van Laar, The Fortran simulation translator, a simulation language, European Journal of Agronomy, Volume 18, Issues 3–4, January 2003, Pages 359-361, ISSN 1161-0301, http://dx.doi.org/10.1016/S1161-0301(02)00131-4.
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
